Solve day 1 part 2
I reconsidered how to solve part 1 after having done this, so I could
use the same strategy for both. Rather than relying on pattern matching
with a custom reduce-type function, I now create the pairs I want to
compare in advance, then reduce against these based on whether the
pairings are equal or not.

I don't love it, but it works, and I have some generalisation by taking
this path.