Sovle d5p2
Initially I implemented this more or less the same as with part 1, using
List for the input. Sadly this took ages (minutes!!) to run, likely due
to having to create a new large list many many times, and that being
very slow.

I swapped to using a map instead for this second part, and this is much
much faster.